vscode

+++ Plotting/rtplot.cpp
@@ -166,11 +166,12 @@ void rtplot::drawData()
         for(size_t i = 0; i < valid_data.size(); i++) {
             const std::vector<FLOAT>* d = (const std::vector<FLOAT>*) (valid_data[i].getData());
             for(auto it = d->begin(); it != d->end(); ++it) {
-                size_t tx = (*it - x_axis.min_value)*width_per_value;
-                size_t ty = (*it - y_axis.min_value)*height_per_value;
+                double tx = (*it - x_axis.min_value)*width_per_value;
+                double ty = (*it - y_axis.min_value)*height_per_value;
                 if(principal_axis == 'x') tx = delta * i + delta/2;
                 if(principal_axis == 'y') ty = delta * i + delta/2;
                 Location tl = origin_plot(tx,ty);
+                //Simd::DrawingLine(m_plot_view,origin_plot(tx,ty+delta/2-1),origin_plot(tx,ty-delta/2+1),Color(0,0,255));
                 Simd::DrawRectangle(m_plot_view,tl.x-1,tl.y-1,tl.x+1,tl.y+1, Color(0,0,255));
             }
         }
